BROKEN! New text rendering WIP 3

+++ cegui/include/CEGUI/falagard/Enums.h
@@ -27,7 +27,8 @@
 #ifndef _CEGUIFalEnums_h_
 #define _CEGUIFalEnums_h_
 
-// Start of CEGUI namespace section
+#include <cstdint>
+
 namespace CEGUI
 {
     /*!
@@ -53,7 +54,7 @@ namespace CEGUI
     \brief
         Enumeration of possible values to indicate the vertical formatting to be used for an image component.
     */
-    enum class VerticalImageFormatting : int
+    enum class VerticalImageFormatting : uint8_t
     {
         TopAligned,         //!< Top of Image should be aligned with the top of the destination area.
         CentreAligned,      //!< Image should be vertically centred within the destination area.
@@ -66,7 +67,7 @@ namespace CEGUI
     \brief
         Enumeration of possible values to indicate the horizontal formatting to be used for an image component.
     */
-    enum class HorizontalFormatting : int
+    enum class HorizontalFormatting : uint8_t
     {
         LeftAligned,        //!< Left of Image should be aligned with the left of the destination area.
         CentreAligned,      //!< Image should be horizontally centred within the destination area.
@@ -79,7 +80,7 @@ namespace CEGUI
     \brief
         Enumeration of possible values to indicate the vertical formatting to be used for a text component.
     */
-    enum class VerticalTextFormatting : int
+    enum class VerticalTextFormatting : uint8_t
     {
         TopAligned,         //!< Top of text should be aligned with the top of the destination area.
         CentreAligned,      //!< text should be vertically centred within the destination area.
@@ -90,7 +91,7 @@ namespace CEGUI
     \brief
         Enumeration of possible values to indicate the horizontal formatting to be used for a text component.
     */
-    enum class HorizontalTextFormatting : int
+    enum class HorizontalTextFormatting : uint8_t
     {
         LeftAligned,                //!< Left of text should be aligned with the left of the destination area (single line of text only).
         RightAligned,               //!< Right of text should be aligned with the right of the destination area  (single line of text only).
